Add Docker support and refactor test environment setup
This commit introduces Docker and Docker-Compose files for setting up the test environment. It also modifies the conftest.py file to run tests in a headless mode. Also, there is a refactoring of the change_name function in personal_page.py to address a problem with input cleaning. A .gitignore file has also been added to the project to exclude certain files from tracking.